test updates

+++ nars_logic/src/main/java/nars/nal/nal4/Image.java
@@ -22,10 +22,10 @@ abstract public class Image extends Compound {
      */
     public final short relationIndex;
 
-    protected Image(Term[] components, short relationIndex) {
+    protected Image(Term[] components, int relationIndex) {
         super(components);
 
-        this.relationIndex = relationIndex;
+        this.relationIndex = (short) relationIndex;
 
         init(components);
     }
